requirements.txt
aiogram~=3.18.0
translate~=3.6.1
python-dotenv~=1.0.1